interaction_tester/requirements.txt
flask==2.0.1
python-dotenv==0.19.0
selenium==4.1.0
google-generativeai==0.3.0
reportlab==3.6.8
python-json-logger==2.0.7